backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
asyncpg==0.29.0
requests==2.31.0
bcrypt==4.0.1
passlib==1.7.4
python-jose==3.3.0
alembic==1.12.1
numpy==1.24.3
scikit-learn==1.3.2